api/requirements.txt
asgiref==3.4.1
click==8.0.1
dnspython==1.16.0
fastapi==0.68.1
h11==0.12.0
importlib-metadata==4.8.1
Jinja2==3.0.1
pydantic==1.8.2
PyJWT==2.1.0
pymongo[srv]
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.5.0
